Adding some delay in reconnect
Adding some delay in the reconnect process, to see if that improves the
chances of it working.

The new reconnect logic works, but it's not bulletproof. I found that
at one point, the reconnetion to the AP seems to have been made, but
would could not connect to the broker, perhaps because the CC3000
wasn't fully ready. Adding a little delay between the reconnect to the
AP and the connection to the broker may help.