Move delay to try to improve recovery
The cc3000 is not a great chip. It seems to have problems with
stability, and will fall down after a few hours of service. There's
already some logic that attempts to recover from those failures, but
it seems to get stuck. Moving the delay to give the chip some time
before we do anything else.